working version, cleaned up controller
