+++ src/plugin/store.ts
@@ -1,0 +1,3 @@
+const store = {remoteComponents: [], successfulNodes: []};
+
+export default store;
